최적화

+++ src/views/ApartSearch.vue
@@ -1,6 +1,6 @@
 <template>
     <div :class="{ 'change-big': $store.state.mainDiv, 'change-small': !$store.state.mainDiv }">
-        <map-view :apartList="apartList" :apartItem="apartItem" ref="MapView"></map-view>
+        <map-view :apartList="apartList" ref="MapView"></map-view>
         <!--시구동 검색 창 시작-->
         <div class="card col-sm-12" style="min-height: 935px; z-index: 1; position: absolute; top: 0; right: 0; z-index: 1; width: 400px">
             <div class="card-body">
@@ -37,9 +37,9 @@
                 <!--아파트명 검색 창 끝-->
                 <!--검색 결과 정렬 및 관심지역 토글 창 시작-->
                 <div class="row">
-                    <div class="col-8  mb-3 pt-1">
-                        <select class="col form-select-sm m-2" v-model="sortType">
-                            <option value="0"> </option>
+                    <div class="col-6  mb-3 pt-1">
+                        <select class="col form-select-sm mt-2" v-model="sortType" style="margin-left: -5px;">
+                            <option value="0">정렬</option>
                             <option value="1">거래가 많은 순</option>
                             <option value="2">최근 건축 순</option>
                             <option value="3">거래가 높은 순</option>
@@ -47,7 +47,7 @@
                             <option value="5">이름순</option>
                         </select>
                     </div>
-                    <div class="col-4  mb-3 pt-1">
+                    <div class="col-6  mb-3 pt-3">
                         관심지역만 보기
                         <input type="checkbox" v-model="interestToggle" true-value="true" false-value="false"/>
                     </div>
@@ -56,9 +56,9 @@
                 <div class="row leftSide">
                     <div class="card p-0">
                         <ul class="list-group list-group-flush" id="searchResult" >
-    <li class="list-group-item p-1" style="cursor: pointer" v-for="(apart, index) in apartList" :key="index" @click="detailApart(apart.aptCode)">
+    <li class="list-group-item p-1" style="cursor: pointer" v-for="(apart, index) in apartList" :key="index" @click="detailApart(apart.aptCode)" @click.stop >
                             <!--클릭 이벤트와 마우스 이벤트들이 자식 요소들에게도 이벤트가 발생함.-->
-        <div class="card border-0 shadow" @mouseover="itemMouseOver($event, apart)" @mouseout="itemMouseOut">
+        <div class="card border-0 shadow apartItem" @mouseover="itemMouseOver($event, apart)" @mouseout="itemMouseOut">
             <div class="card-body p-1">
                 <div class="row d-block d-xl-flex align-items-center">
                     <div class="col-12 col-xl-3 text-xl-center mb-3 mb-xl-0 d-flex align-items-center justify-content-xl-center">
@@ -83,8 +83,8 @@
                         </div>
                     </div>
                     <div class="col-1 col-xl-1 px-xl-0">
-                        <font-awesome-icon icon="fa-solid fa-heart" style="color:#ca1a41;" v-if="apart.isUserInterest === 1" @click="deleteInterestApart(apart)"/>
-                        <font-awesome-icon icon="fa-regular fa-heart" v-else  @click="insertInterestApart(apart)"/>
+                        <font-awesome-icon v-bind:icon="heartType[apart.isUserInterest]" 
+                            v-bind:style="heartStyle[apart.isUserInterest]"  @click.stop="changeInterestApart(apart)" />
                     </div>
                 </div>
             </div>
@@ -159,6 +159,12 @@ export default {
             //detail modal info
             aptCode: "",
             detailModal: null,
+
+
+            //
+            heartType:[ "fa-regular fa-heart", "fa-solid fa-heart"],
+            heartStyle:[ {}, {color:"#ca1a41"}],
+
         };
     },
     mounted() {
@@ -277,34 +283,35 @@ export default {
             this.detailModal.show();
             //displayMarkers(data);
         },
-        insertInterestApart : async function (apart) {
-            console.log(this.userInfo.userId + " " + apart.aptCode);
-            let response = await http.post("/" + this.userInfo.userId + "/interest/aparts", {userId : this.userInfo.userId, aptCode : apart.aptCode});
+        // 관심 아파트 토글
+        changeInterestApart : async function (apart) {
+            let response = null;
+            if(apart.isUserInterest == 1){
+                response = await http.delete("/" + this.userInfo.userId + "/interest/aparts/" + apart.aptCode);
+            }
+            else{
+                response = await http.post("/" + this.userInfo.userId + "/interest/aparts", {userId : this.userInfo.userId, aptCode : apart.aptCode});
+            }
+            
             let { data } = response;
-            console.log("insertInterestApart() :")
-            console.log(data);
-            //if (data.result == 1)apart.isUserInterest = 1;
-            this.getSearchResult();
-            //this.sidoList = data;
-        },
-        deleteInterestApart : async function (apart) {
-            let response = await http.delete("/" + this.userInfo.userId + "/interest/aparts/" + apart.aptCode);
-            let { data } = response;
-            console.log("deleteInterestApart() :")
-            console.log(data);
-            //if (data.result == 1) apart.isUserInterest = 0;
-            this.getSearchResult();
-            //this.sidoList = data;
-        },
 
+            if (data == 1) {
+                apart.isUserInterest = (apart.isUserInterest+1) % 2;
+                if(this.interestToggle){
+                    this.getSearchResult();
+                }
+            }
+            //this.getSearchResult();
+            //this.sidoList = data;
+        },
         //검색 결과 리스트 이벤트 등록
         itemMouseOver : function(event, apart){
-            event.target.style.background = "gray";
+            //event.target.style.background = "gray";
             //console.log(apart)
             this.apartItem = apart
         },
         itemMouseOut : function(event){
-            event.target.style.background = "white";
+            //event.target.style.background = "white";
             this.apartItem = null;
         },
         
@@ -323,6 +330,9 @@ export default {
     display: none;
 }
 
+.apartItem:hover{
+    background: lightgrey;
+}
 /* .list-group-item:nth-child(2n) {
     background-color: lightgray;
 } */
